backend/requirements.txt
flask==3.0.3
mongoengine==0.29.1
flask-mongoengine-3==1.1.0
werkzeug==3.0.3
python-dotenv==1.2.1